Moved authorization helper to authenticable concern
