+++ config/locales/en.yml
@@ -36,6 +36,8 @@ en:
   vote:
     success_message: "Thanks for voting!"
     duplicate_error_message: "You have already voted!"
+  authorization:
+    denied: "Access denied. You are not authorized to perform this action!"
   successfully_created: "%{entity} was successfully created!"
   successfully_updated: "%{entity} was successfully updated!"
   successfully_deleted: "%{entity} was successfully deleted!"
